+++ node/src/effect/announcements.rs
@@ -17,6 +17,7 @@ use casper_types::{EraId, ExecutionEffect, ExecutionResult, PublicKey, Timestamp
 use crate::{
     components::{
         chainspec_loader::NextUpgrade, deploy_acceptor::Error, diagnostics_port::FileSerializer,
+        small_network::blocklist::BlocklistJustification,
     },
     effect::Responder,
     types::{
@@ -225,14 +226,22 @@ impl Display for ConsensusAnnouncement {
 #[derive(Debug, Serialize)]
 pub(crate) enum BlocklistAnnouncement {
     /// A given peer committed a blockable offense.
-    OffenseCommitted(Box<NodeId>),
+    OffenseCommitted {
+        /// The peer ID of the offending node.
+        offender: Box<NodeId>,
+        /// Justification for blocking the peer.
+        justification: Box<BlocklistJustification>,
+    },
 }
 
 impl Display for BlocklistAnnouncement {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match self {
-            BlocklistAnnouncement::OffenseCommitted(peer) => {
-                write!(f, "peer {} committed offense", peer)
+            BlocklistAnnouncement::OffenseCommitted {
+                offender,
+                justification,
+            } => {
+                write!(f, "peer {} committed offense: {}", offender, justification)
             }
         }
     }