+++ node/CHANGELOG.md
@@ -10,6 +10,15 @@ All notable changes to this project will be documented in this file.  The format
 [comment]: <> (Security:   in case of vulnerabilities)
 
 
+## [Unreleased (next)]
+
+### Added
+
+* Add `net-info` command to diagnostics port to gain insights into the nodes network connectivity state.
+
+### Changed
+
+* Log messages for blocked nodes have been unified and reasons for blocking peers are better tracked.
 
 ## [Unreleased]